stinky numpy

+++ spinn_front_end_common/utility_models/reverse_ip_tag_multicast_source_machine_vertex.py
@@ -272,7 +272,8 @@ class ReverseIPTagMulticastSourceMachineVertex(
         :rtype: int
         """
         if len(send_buffer_times) and hasattr(send_buffer_times[0], "__len__"):
-            counts = numpy.bincount(numpy.concatenate(send_buffer_times))
+            counts = numpy.bincount(
+                numpy.concatenate(send_buffer_times).astype(numpy.int64))
             if len(counts):
                 return max(counts)
             return 0
